Moved controls to their own folder, added a render to each of them and made them use the actual cursor to print.
